Update RSpec syntax for testing truthiness
The `be_true` and `be_false` matchers have been deprecated many versions
ago in favour of `be_truthy` and `be_falsey` or (even better) `be true`
and `be false`. [1]

The `be_true` and `be_false` matchers were completely removed in a more
recent version. [2] This was causing the tests to fail.

1: https://github.com/rspec/rspec-expectations/blob/master/Changelog.md#2990beta1--2013-11-07
2: https://github.com/rspec/rspec-expectations/blob/master/Changelog.md#300beta1--2013-11-07